prevent trustees from being removed from the admin group

+++ app/controllers/RoleUsersController.php
@@ -45,6 +45,15 @@ class RoleUsersController extends \BaseController {
 	public function destroy($roleId, $userId)
 	{
         $role = Role::findOrFail($roleId);
+
+        //don't let people remove the admin permission if they are a trustee
+
+        $user = User::findOrFail($userId);
+        if ($user->director && $role->name == 'admin') {
+            Notification::error("You cannot remove a trustee from the admin group");
+            return Redirect::back();
+        }
+
         $role->users()->detach($userId);
 
         return Redirect::back();
